// packages/subgraph/src/types/nftx-v3-uniswap.d.ts
export type Maybe<T> = T | null;
// ...
export type Account = {
  __typename?: 'Account';
  /**  Number of closed positions this account has  */
  closedPositionCount: Scalars['Int']['output'];
  /**  Number of deposits this account made  */
  depositCount: Scalars['Int']['output'];
  /**  All deposit events of this account  */
  deposits: Array<Deposit>;
  /**  { Account address }  */
  id: Scalars['Bytes']['output'];
  /**  Number of open positions this account has  */
  openPositionCount: Scalars['Int']['output'];
  /**  Number of positions this account has  */
  positionCount: Scalars['Int']['output'];
  /**  All positions that belong to this account  */
  positions: Array<Position>;
  /**  Number of times this account has traded/swapped */
  swapCount: Scalars['Int']['output'];
  /**  All swap events of this account  */
  swaps: Array<Swap>;
  /**  Number of withdrawals this account made  */
  withdrawCount: Scalars['Int']['output'];
  /**  All withdraw events of this account  */
  withdraws: Array<Withdraw>;
};
// ...
export type ActiveAccount = {
  __typename?: 'ActiveAccount';
  /**  { daily/hourly }-{ Address of the account }-{ Days/hours since Unix epoch }  */
  id: Scalars['Bytes']['output'];
};
// ...
export type Deposit = {
  __typename?: 'Deposit';
  /**  Account that emitted this event  */
  account: Account;
  /**  USD-normalized value of the transaction of the underlying (e.g. sum of tokens deposited into a pool)  */
  amountUSD: Scalars['BigDecimal']['output'];
  /**  Block number of this event  */
  blockNumber: Scalars['BigInt']['output'];
  /**  Gas limit of the transaction that emitted this event  */
  gasLimit?: Maybe<Scalars['BigInt']['output']>;
  /**  Gas price of the transaction that emitted this event  */
  gasPrice?: Maybe<Scalars['BigInt']['output']>;
  /**  Gas used in this transaction. (Optional because not every chain will support this)  */
  gasUsed?: Maybe<Scalars['BigInt']['output']>;
  /**  Transaction hash of the transaction that emitted this event  */
  hash: Scalars['Bytes']['output'];
  /**  { Transaction hash }-{ Log index }  */
  id: Scalars['Bytes']['output'];
  /**  Amount of input tokens in the token's native unit  */
  inputTokenAmounts: Array<Scalars['BigInt']['output']>;
  /**  Input tokens of the pool. E.g. WETH and USDC to a WETH-USDC pool  */
  inputTokens: Array<Token>;
  /**  Amount of liquidity minted  */
  liquidity: Scalars['BigInt']['output'];
  /**  Event log index. For transactions that don't emit event, create arbitrary index starting from 0  */
  logIndex: Scalars['Int']['output'];
  /**  Nonce of the transaction that emitted this event  */
  nonce: Scalars['BigInt']['output'];
  /**  The pool involving this event  */
  pool: LiquidityPool;
  /**  The user position changed by this event  */
  position?: Maybe<Position>;
  /**  The protocol this transaction belongs to  */
  protocol: DexAmmProtocol;
  /**  Amount of input tokens in the liquidity pool  */
  reserveAmounts?: Maybe<Array<Scalars['BigInt']['output']>>;
  /**  lower tick of position  */
  tickLower?: Maybe<Scalars['BigInt']['output']>;
  /**  upper tick of position  */
  tickUpper?: Maybe<Scalars['BigInt']['output']>;
  /**  Timestamp of this event  */
  timestamp: Scalars['BigInt']['output'];
};
// ...
export type DexAmmProtocol = Protocol & {
  __typename?: 'DexAmmProtocol';
  /**  This is a boolean to indicate whether or not the pools have been instantiated the were initialized before Optimism regenesis  */
  _regenesis: Scalars['Boolean']['output'];
  /**  All liquidity in USD that is active. Will be equal to totalLiquidity except for in concentrated liquidity - where activeLiquidity is all liquidity positions that contain the pools current tick.  */
  activeLiquidityUSD: Scalars['BigDecimal']['output'];
  /**  Total number of positions (open and closed)  */
  cumulativePositionCount: Scalars['Int']['output'];
  /**  Gross revenue for the protocol (revenue claimed by protocol). Examples: AMM protocol fee (Sushi’s 0.05%). OpenSea 10% sell fee.  */
  cumulativeProtocolSideRevenueUSD: Scalars['BigDecimal']['output'];
  /**  Revenue claimed by suppliers to the protocol. LPs on DEXs (e.g. 0.25% of the swap fee in Sushiswap). Depositors on Lending Protocols. NFT sellers on OpenSea.  */
  cumulativeSupplySideRevenueUSD: Scalars['BigDecimal']['output'];
  /**  All revenue generated by the protocol. e.g. 0.30% of swap fee in Sushiswap, all yield generated by Yearn.  */
  cumulativeTotalRevenueUSD: Scalars['BigDecimal']['output'];
  /**  Number of cumulative liquidity providers  */
  cumulativeUniqueLPs: Scalars['Int']['output'];
  /**  Number of cumulative traders  */
  cumulativeUniqueTraders: Scalars['Int']['output'];
  /**  Number of cumulative unique users  */
  cumulativeUniqueUsers: Scalars['Int']['output'];
  /**  All historical volume in USD  */
  cumulativeVolumeUSD: Scalars['BigDecimal']['output'];
  /**  Daily usage metrics for this protocol  */
  dailyUsageMetrics: Array<UsageMetricsDailySnapshot>;
  /**  Daily financial metrics for this protocol  */
  financialMetrics: Array<FinancialsDailySnapshot>;
  /**  Hourly usage metrics for this protocol  */
  hourlyUsageMetrics: Array<UsageMetricsHourlySnapshot>;
  /**  Smart contract address of the protocol's main contract (Factory, Registry, etc)  */
  id: Scalars['Bytes']['output'];
  /**  Day ID of the most recent daily snapshot  */
  lastSnapshotDayID: Scalars['Int']['output'];
  /**  Block number of the last time this entity was updated  */
  lastUpdateBlockNumber: Scalars['BigInt']['output'];
  /**  Timestamp of the last time this entity was updated  */
  lastUpdateTimestamp: Scalars['BigInt']['output'];
  /**  Version of the methodology used to compute metrics, loosely based on SemVer format (e.g. 1.0.0)  */
  methodologyVersion: Scalars['String']['output'];
  /**  Name of the protocol, including version. e.g. Uniswap v3  */
  name: Scalars['String']['output'];
  /**  The blockchain network this subgraph is indexing on  */
  network: Network;
  /**  Total number of open positions  */
  openPositionCount: Scalars['Int']['output'];
  /**  All pools that belong to this protocol  */
  pools: Array<LiquidityPool>;
  /**  Current PCV (Protocol Controlled Value). Only relevant for protocols with PCV.  */
  protocolControlledValueUSD?: Maybe<Scalars['BigDecimal']['output']>;
  /**  Version of the subgraph schema, in SemVer format (e.g. 1.0.0)  */
  schemaVersion: Scalars['String']['output'];
  /**  Slug of protocol, including version. e.g. uniswap-v3  */
  slug: Scalars['String']['output'];
  /**  Version of the subgraph implementation, in SemVer format (e.g. 1.0.0)  */
  subgraphVersion: Scalars['String']['output'];
  /**  The sum of all active and non-active liquidity in USD for this pool.  */
  totalLiquidityUSD: Scalars['BigDecimal']['output'];
  /**  Total number of pools  */
  totalPoolCount: Scalars['Int']['output'];
  /**  Current TVL (Total Value Locked) of the entire protocol  */
  totalValueLockedUSD: Scalars['BigDecimal']['output'];
  /**  The type of protocol (e.g. DEX, Lending, Yield, etc)  */
  type: ProtocolType;
  /**  All protocol-side value locking in USD that remains uncollected and unused in the protocol.  */
  uncollectedProtocolSideValueUSD: Scalars['BigDecimal']['output'];
  /**  All supply-side value locking in USD that remains uncollected and unused in the protocol.  */
  uncollectedSupplySideValueUSD: Scalars['BigDecimal']['output'];
};
// ...
export type LiquidityPool = {
// ...
  id: Scalars['Bytes']['output'];
// ...
};
// ...
export type LiquidityPoolFee = {
  __typename?: 'LiquidityPoolFee';
  /**  Fee as a percentage of the trade (swap) amount. Does not always apply  */
  feePercentage?: Maybe<Scalars['BigDecimal']['output']>;
  /**  Type of fee this pool uses  */
  feeType: LiquidityPoolFeeType;
  /**  { Fee type }-{ Pool address }  */
  id: Scalars['Bytes']['output'];
};
// ...
export enum ProtocolType {
  Bridge = 'BRIDGE',
  Exchange = 'EXCHANGE',
  Generic = 'GENERIC',
  Lending = 'LENDING',
  Yield = 'YIELD'
}
// ...
export type Query = {
  __typename?: 'Query';
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  account?: Maybe<Account>;
  accounts: Array<Account>;
  activeAccount?: Maybe<ActiveAccount>;
  activeAccounts: Array<ActiveAccount>;
  deposit?: Maybe<Deposit>;
  deposits: Array<Deposit>;
  dexAmmProtocol?: Maybe<DexAmmProtocol>;
  dexAmmProtocols: Array<DexAmmProtocol>;
  financialsDailySnapshot?: Maybe<FinancialsDailySnapshot>;
  financialsDailySnapshots: Array<FinancialsDailySnapshot>;
  helperStore?: Maybe<_HelperStore>;
  helperStores: Array<_HelperStore>;
  liquidityPool?: Maybe<LiquidityPool>;
  liquidityPoolAmount?: Maybe<_LiquidityPoolAmount>;
  liquidityPoolAmounts: Array<_LiquidityPoolAmount>;
  liquidityPoolDailySnapshot?: Maybe<LiquidityPoolDailySnapshot>;
  liquidityPoolDailySnapshots: Array<LiquidityPoolDailySnapshot>;
  liquidityPoolFee?: Maybe<LiquidityPoolFee>;
  liquidityPoolFees: Array<LiquidityPoolFee>;
  liquidityPoolHourlySnapshot?: Maybe<LiquidityPoolHourlySnapshot>;
  liquidityPoolHourlySnapshots: Array<LiquidityPoolHourlySnapshot>;
  liquidityPools: Array<LiquidityPool>;
  position?: Maybe<Position>;
  positionSnapshot?: Maybe<PositionSnapshot>;
  positionSnapshots: Array<PositionSnapshot>;
  positions: Array<Position>;
  protocol?: Maybe<Protocol>;
  protocols: Array<Protocol>;
  rewardToken?: Maybe<RewardToken>;
  rewardTokens: Array<RewardToken>;
  swap?: Maybe<Swap>;
  swaps: Array<Swap>;
  tick?: Maybe<Tick>;
  tickDailySnapshot?: Maybe<TickDailySnapshot>;
  tickDailySnapshots: Array<TickDailySnapshot>;
  tickHourlySnapshot?: Maybe<TickHourlySnapshot>;
  tickHourlySnapshots: Array<TickHourlySnapshot>;
  ticks: Array<Tick>;
  token?: Maybe<Token>;
  tokenWhitelist?: Maybe<_TokenWhitelist>;
  tokenWhitelistSymbol?: Maybe<_TokenWhitelistSymbol>;
  tokenWhitelistSymbols: Array<_TokenWhitelistSymbol>;
  tokenWhitelists: Array<_TokenWhitelist>;
  tokens: Array<Token>;
  usageMetricsDailySnapshot?: Maybe<UsageMetricsDailySnapshot>;
  usageMetricsDailySnapshots: Array<UsageMetricsDailySnapshot>;
  usageMetricsHourlySnapshot?: Maybe<UsageMetricsHourlySnapshot>;
  usageMetricsHourlySnapshots: Array<UsageMetricsHourlySnapshot>;
  withdraw?: Maybe<Withdraw>;
  withdraws: Array<Withdraw>;
};
// ...
export type RewardToken = {
  __typename?: 'RewardToken';
  /**  { Reward token type }-{ Smart contract address of the reward token }  */
  id: Scalars['Bytes']['output'];
  /**  Reference to the actual token  */
  token: Token;
  /**  The type of the reward token  */
  type: RewardTokenType;
};
// ...
export type _HelperStore = {
  __typename?: '_HelperStore';
  id: Scalars['Bytes']['output'];
  /**  price of ETH/TVL in USD  */
  valueDecimal?: Maybe<Scalars['BigDecimal']['output']>;
  /**  Token Prices  */
  valueDecimalList?: Maybe<Array<Scalars['BigDecimal']['output']>>;
  /**  # of deposits, # of unique users  */
  valueInt?: Maybe<Scalars['Int']['output']>;
};
// ...
export type _LiquidityPoolAmount = {
  __typename?: '_LiquidityPoolAmount';
  /**  Smart contract address of the pool  */
  id: Scalars['Bytes']['output'];
  /**  Amount of input tokens in the pool. The ordering should be the same as the pool's `inputTokens` field.  */
  inputTokenBalances: Array<Scalars['BigDecimal']['output']>;
  /**  Input tokens of the pool (not input tokens of the event/transaction). E.g. WETH and USDC from a WETH-USDC pool  */
  inputTokens: Array<Token>;
  tokenPrices: Array<Scalars['BigDecimal']['output']>;
};
// ...
export type _Meta_ = {
  __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars['String']['output'];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars['Boolean']['output'];
};